backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
opencv-python==4.8.1.78
pytesseract==0.3.10
matplotlib==3.8.2
spacy==3.7.2
pillow==10.1.0
numpy==1.25.2
scikit-learn==1.3.2
pandas==2.1.4
aiofiles==23.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
